+++ cmd/stunstamp/stunstamp.go
@@ -34,10 +34,10 @@ import (
 	"github.com/golang/snappy"
 	"github.com/prometheus/prometheus/prompb"
 	"github.com/tcnksm/go-httpstat"
-	"tailscale.com/logtail/backoff"
 	"tailscale.com/net/stun"
 	"tailscale.com/net/tcpinfo"
 	"tailscale.com/tailcfg"
+	"tailscale.com/util/backoff"
 )
 
 var (
